+++ packages/cli-driver/spec/critical/writeSpec.ts
@@ -23,13 +23,13 @@ describe('write', () => {
 
     let t1: any = time()
     await client.enter('echo 123')
-    let data: any = await client.waitForData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
+    let data: any = await client.forData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
     expect(data.type).not.toBe(Driver.ERROR_TYPE)
     t1 = time(t1)
 
     let t2: any = time()
     await client.enter('echo 123', 500) // notice the second number
-    data = await client.waitForData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
+    data = await client.forData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
     expect(data.type).not.toBe(Driver.ERROR_TYPE)
     t2 = time(t2)
 
@@ -57,13 +57,13 @@ describe('write', () => {
 
     let t1: any = time()
     await client.enter('echo 123', 20) // we set locally a much lesser value for waitAfterEnter it should override the global one)
-    let data: any = await client.waitForData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
+    let data: any = await client.forData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
     expect(data.type).not.toBe(Driver.ERROR_TYPE)
     t1 = time(t1)
 
     let t2: any = time()
     await client.enter('echo 123') // waitAfterEnter was set globally!
-    data = await client.waitForData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
+    data = await client.forData({ predicate: '123', interval: 20 , timeout: 600, rejectOnTimeout: false })
     expect(data.type).not.toBe(Driver.ERROR_TYPE)
     t2 = time(t2)